Move fanoutconsumer into consumer package.
