+++ .chloggen/connectors-routing-v3.yaml
@@ -8,7 +8,7 @@ component: connectors
 note: Provide connectors with a mechanism to route data to specific pipelines
 
 # One or more tracking issues or pull requests related to the change
-issues: [7152]
+issues: [7179]
 
 # (Optional) One or more lines of additional information to render under the primary note.
 # These lines will be padded with 2 spaces and then inserted directly into the document.
