+++ consumer/fanoutconsumer/logs_test.go
@@ -17,17 +17,13 @@ package fanoutconsumer
 import (
 	"context"
 	"errors"
-	"fmt"
-	"strconv"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 
-	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/consumer/consumertest"
 	"go.opentelemetry.io/collector/internal/testdata"
-	"go.opentelemetry.io/collector/pdata/plog"
 )
 
 func TestLogsNotMultiplexing(t *testing.T) {
@@ -204,114 +200,4 @@ type mutatingErr struct {
 
 func (mts mutatingErr) Capabilities() consumer.Capabilities {
 	return consumer.Capabilities{MutatesData: true}
-}
-
-func TestLogsRouterMultiplexing(t *testing.T) {
-	var max = 20
-	for numIDs := 1; numIDs < max; numIDs++ {
-		for numCons := 1; numCons < max; numCons++ {
-			for numLogs := 1; numLogs < max; numLogs++ {
-				t.Run(
-					fmt.Sprintf("%d-ids/%d-cons/%d-logs", numIDs, numCons, numLogs),
-					fuzzLogsRouter(numIDs, numCons, numLogs),
-				)
-			}
-		}
-	}
-}
-
-func fuzzLogsRouter(numIDs, numCons, numLogs int) func(*testing.T) {
-	return func(t *testing.T) {
-		allIDs := make([]component.ID, 0, numCons)
-		allCons := make([]consumer.Logs, 0, numCons)
-		allConsMap := make(map[component.ID]consumer.Logs)
-
-		// If any consumer is mutating, the router must report mutating
-		for i := 0; i < numCons; i++ {
-			allIDs = append(allIDs, component.NewIDWithName("sink", strconv.Itoa(numCons)))
-			// Random chance for each consumer to be mutating
-			if (numCons+numLogs+i)%4 == 0 {
-				allCons = append(allCons, &mutatingLogsSink{LogsSink: new(consumertest.LogsSink)})
-			} else {
-				allCons = append(allCons, new(consumertest.LogsSink))
-			}
-			allConsMap[allIDs[i]] = allCons[i]
-		}
-
-		r := NewLogsRouter(allConsMap)
-		assert.False(t, r.Capabilities().MutatesData)
-
-		consumers := r.Consumers()
-
-		ld := testdata.GenerateLogs(1)
-
-		// Keep track of how many logs each consumer should receive.
-		// This will be validated after every call to RouteLogs.
-		expected := make(map[component.ID]int, numCons)
-
-		for i := 0; i < numLogs; i++ {
-			// Build a random set of ids (no duplicates)
-			randCons := make(map[component.ID]bool, numIDs)
-			for j := 0; j < numIDs; j++ {
-				// This number should be pretty random and less than numCons
-				conNum := (numCons + numIDs + i + j) % numCons
-				randCons[allIDs[conNum]] = true
-			}
-
-			lcs := make([]consumer.Logs, 0, len(randCons))
-			for id := range randCons {
-				lcs = append(lcs, consumers[id])
-				expected[id]++
-			}
-
-			// Route to list of consumers
-			assert.NoError(t, NewLogs(lcs).ConsumeLogs(context.Background(), ld))
-
-			// Validate expectations for all consumers
-			for id := range expected {
-				logs := []plog.Logs{}
-				switch con := allConsMap[id].(type) {
-				case *consumertest.LogsSink:
-					logs = con.AllLogs()
-				case *mutatingLogsSink:
-					logs = con.AllLogs()
-				}
-				assert.Len(t, logs, expected[id])
-				for n := 0; n < len(logs); n++ {
-					assert.EqualValues(t, ld, logs[n])
-				}
-			}
-		}
-	}
-}
-
-func TestLogsRouterGetConsumer(t *testing.T) {
-	ctx := context.Background()
-	ld := testdata.GenerateLogs(1)
-
-	fooID := component.NewID("foo")
-	barID := component.NewID("bar")
-
-	foo := new(consumertest.LogsSink)
-	bar := new(consumertest.LogsSink)
-	r := NewLogsRouter(map[component.ID]consumer.Logs{fooID: foo, barID: bar})
-	assert.Len(t, foo.AllLogs(), 0)
-	assert.Len(t, bar.AllLogs(), 0)
-
-	cons := r.Consumers()
-
-	both := NewLogs([]consumer.Logs{cons[fooID], cons[barID]})
-	assert.NoError(t, both.ConsumeLogs(ctx, ld))
-	assert.Len(t, foo.AllLogs(), 1)
-	assert.Len(t, bar.AllLogs(), 1)
-
-	fooOnly := NewLogs([]consumer.Logs{cons[fooID]})
-	assert.NoError(t, fooOnly.ConsumeLogs(ctx, ld))
-	assert.Len(t, foo.AllLogs(), 2)
-	assert.Len(t, bar.AllLogs(), 1)
-
-	barOnly := NewLogs([]consumer.Logs{cons[barID]})
-	assert.NoError(t, barOnly.ConsumeLogs(ctx, ld))
-	assert.Len(t, foo.AllLogs(), 2)
-	assert.Len(t, bar.AllLogs(), 2)
 }
